Move code for fetching assets into own type
